feat: Route | RouterLink | paraMap | Display.
It's defined that any url which begins with /book-flight/ and also has a parameter will load 'BookflightComponent'.

We use 'routerLink' to build the url dinamically by taking 'flight.id'.

We use the paramMap from 'ActivatedRoute' to get access to the route information, this case the 'flightId'.

Finally, the flightId gets displayed on screen if it exists.