feat: Domain validation | From immutable to mutable | Catching error in client side | Getting rid of Guid.
Domain validation was added for remaining number of seats as its decrement as well.

The 'RemainingNumberOfSeats' property was changed to mutable since we needed to decrement this number, this was due to it's a record.

Catching the 409 error to show it to the user.

The FlightId porperty is no longer needed since the Booking it's attached to the Flight.