refactor: EF |Entities into DBContext | Program.cs modifications | Flight class updated.
EF package was added to the project

The Entities clasas was turned into a DBContext by adding DBSet for tables instead if List/IList .

The main program file was updated to add the DBContext into the DI container and also SaveChanges method was added to persist the in-memory DB.

The Flight record was turned into a class since we cannot persist a record into a DB context if the recordd changes as Flight does. Unlike Passenger, which does not change.