feat: BookingRm | List() method.
The BookingRm record was created to be the template to return to the client while fetching data.

A constructor was added into the BookingController to perform the DI for the DBContext, as the List() method was added as well to return the List of bookings for a certain user email.