feat: BookingController | Flight class | NotFoundError class.
Cancel() method was added to find the flight, make a try to cancel the booking, if something went wrong, an exception will be thrown, if not, the booking cancel is persistent and NoContent() is returned.

CancelBooking() was implemented to look up the booking and if it doesn't exists an exception is thrown, if not the booking changes its state in memory to removed and we make the increment on the number of seats.